Update SDK to support Intear Wallet
Update Godot to 3.6.1
Rework the SDK and test scene to support Intear Wallet, using the desktop wallet if running from desktop, and the web wallet if running from web export. The SDK now currently supports connecting wallet, and signing and sending transactions if function call key was added by user on sign-in.